Almost entirely functional in 2 commits, I feel bad :(
I don't usually do this, but this is the second commit and I already
have a stable working base which is mainly copied from mtop. I mainly
as I cut a lot of stuff out. The eventual core function is however still
missing though, but that should get fixed in future commits.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>