Fixed the disconnecting too early bug
The disconnects were too fast sometimes with as result that the
other connection (could be either the client or the proxied connection)
could miss some data. For example, client sends 100 bytes to multiproto
while multiproto was passing this on to the proxied connection the client
disconnects at which the client connection would get disconnected right
away as well. With as a possible result that the proxy has not yet received
all these 100 bytes.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>